Renamed from ReactNativeBoilerplate
- Update displayName in app.json to the new name
- Delete ios/ and android/ directories
- Run `react-native eject`
- Run `react-native link`
- Rename Registered app component in index.ios.js and index.android.js
