+++ android/app/src/main/java/com/dora/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnativeboilerplate;
+package com.dora;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "ReactNativeBoilerplate";
+        return "Dora";
     }
 }
