+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'ReactNativeBoilerplate'
+rootProject.name = 'Dora'
 
 include ':app'
