+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ReactNativeBoilerplate",
-  "displayName": "ReactNativeBoilerplate"
+  "name": "Dora",
+  "displayName": "Dora"
 }
