+++ ios/Dora.xcodeproj/xcshareddata/xcschemes/Dora-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBoilerplate-tvOS.app"
-               BlueprintName = "ReactNativeBoilerplate-tvOS"
-               ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+               BuildableName = "Dora-tvOS.app"
+               BlueprintName = "Dora-tvOS"
+               ReferencedContainer = "container:Dora.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBoilerplate-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBoilerplate-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+               BuildableName = "Dora-tvOSTests.xctest"
+               BlueprintName = "Dora-tvOSTests"
+               ReferencedContainer = "container:Dora.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBoilerplate-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBoilerplate-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+               BuildableName = "Dora-tvOSTests.xctest"
+               BlueprintName = "Dora-tvOSTests"
+               ReferencedContainer = "container:Dora.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBoilerplate-tvOS.app"
-            BlueprintName = "ReactNativeBoilerplate-tvOS"
-            ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+            BuildableName = "Dora-tvOS.app"
+            BlueprintName = "Dora-tvOS"
+            ReferencedContainer = "container:Dora.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBoilerplate-tvOS.app"
-            BlueprintName = "ReactNativeBoilerplate-tvOS"
-            ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+            BuildableName = "Dora-tvOS.app"
+            BlueprintName = "Dora-tvOS"
+            ReferencedContainer = "container:Dora.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBoilerplate-tvOS.app"
-            BlueprintName = "ReactNativeBoilerplate-tvOS"
-            ReferencedContainer = "container:ReactNativeBoilerplate.xcodeproj">
+            BuildableName = "Dora-tvOS.app"
+            BlueprintName = "Dora-tvOS"
+            ReferencedContainer = "container:Dora.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
