+++ ios/DoraTests/DoraTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface ReactNativeBoilerplateTests : XCTestCase
+@interface DoraTests : XCTestCase
 
 @end
 
-@implementation ReactNativeBoilerplateTests
+@implementation DoraTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
