+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ReactNativeBoilerplate",
+  "name": "Dora",
   "version": "0.0.1",
   "private": true,
   "scripts": {
